requirements.txt
branca==0.8.2
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
folium==0.20.0
fonttools==4.60.1
googlemaps==4.10.0
idna==3.11
jinja2==3.1.6
kiwisolver==1.4.9
markupsafe==3.0.3
matplotlib==3.10.7
narwhals==2.8.0
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.3.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
seaborn==0.13.2
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
wordcloud==1.9.4
xyzservices==2025.4.0
